fix(viewer): address PR #588 second-round CodeRabbit feedback (12 items)
Twelve findings from CodeRabbit's review of the merged Filter
migration. All addressed.

File-size compliance (AGENTS.md §7)
  - SearchModal.filter.builder.tsx: 838 → 187 lines (orchestrator
    only). Toolbar, per-kind editors, and shared widgets extracted to
    filter/{toolbar,rule-editors,widgets}.tsx.
  - SearchModal.filter.tsx: 526 → 393 lines. Result table, error
    box, and rule summary moved to filter/result.tsx.
  - All split modules under 350 lines.

isFilterRule shape validation (Major)
  Per-kind narrowing inside the type guard so a tampered preset like
  `{kind:'ifcType'}` (no values) fails at the deserialise boundary
  instead of crashing the evaluator at `rule.values.some(...)`.
  String-array / op-enum / numeric-finite checks added; quantity
  values must be finite numbers, property values must be strings.
  New tests cover every kind's failure modes.

discoverPropertyAndQuantitySchema sample cap (Major)
  AGENTS.md §2 violation — `extractPropertiesOnDemand` was being
  called for every entity in `onDemandPropertyMap`. New
  SCHEMA_DISCOVERY_SAMPLE_LIMIT=200 caps the walk; in practice every
  distinct (set, prop) pair is captured by the first ~200 entities
  because pset shapes repeat across entities of the same type.

Truncation signal (Minor)
  evaluateFilterRulesFederated now returns
  `{ matches, truncated, scanned }` instead of `FilteredElement[]`.
  `truncated` is true only when the loop stopped because `out.length
  === limit` AND there were unscanned entries — distinguishing "exact
  cap with no remaining work" (NOT truncation) from "stopped
  mid-scan" (IS truncation). UI's "limited to N" badge uses this.
  Side-fix in the chunk loop: `scanned` was credited per-chunk, so a
  break on limit hid truncation; now incremented per inner iteration.

Multi-model banner count (Minor)
  Used `models.size` which counts every entry including those still
  parsing or metadata-only. Changed to count only models with a
  populated `ifcDataStore` — what the evaluator can actually walk.

Stale model id pruning (Minor)
  toggleSearchModelFilter wasn't pruning ids that had left the
  federation, so the persisted Set accumulated garbage across model
  load/unload cycles. Prune `next` against `availableModelIds`
  before computing the "all included" collapse.

Numeric input clearing (Minor x2)
  `setFilterLimit(parseInt(empty)||0)` snapped the field to 0 when
  the user backspaced, blocking edits like 500→50. New
  `<NumericInput>` widget holds an in-flight string buffer locally;
  parent only sees a parsed number when valid. Limit input + quantity
  value input both wired through it. On blur, invalid buffers snap
  back to the last committed value.

CSV CRLF (Minor)
  RFC 4180 specifies CRLF line endings; formatCsv was emitting LF.
  Excel's "Get Data" import on Windows balks at LF-only. Switched
  to CRLF + updated tests.

Bare catch documentation (Minor)
  Both saved-filters.ts and recent-searches.ts had bare `catch {}`
  in the localStorage probe. Per AGENTS.md §7, swallowed exceptions
  need an inline comment; added one explaining the intent (private
  browsing / sandboxed iframe / Safari quota probe failure → silent
  degrade rather than crash).

Doc comment fix (Minor)
  evaluateFilterRulesFederated's docstring claimed results were
  "sorted"; they aren't. Updated to describe the actual ordering
  (model order then per-model iteration order) and note that
  callers should sort if needed.

Test status
  - 385/385 search + slice tests.
  - 645/646 full viewer suite (1 skipped, 0 fail).
  - 26/26 ESM smoke.
  - Production build green.
  - Type-check clean on every touched file.
